added sequelize
