+++ chatroom/.env
@@ -1,0 +1,1 @@
+MYSQL_KEY=Il0ve0scar
