+++ chatroom/schema.sql
@@ -1,0 +1,2 @@
+DROP DATABASE IF EXISTS `sequelize_chatroom_db`;
+CREATE DATABASE `sequelize_chatroom_db`;
